double points

+++ lib/dragbutton.dart
@@ -69,10 +69,8 @@ class _DragButtonState extends State<DragButton> {
         onPanUpdate: (details) {
           setState(() {
             currentPosition = Point.parameterized(
-                // x: (details.localPosition.dx).toInt(),
-                // y: (details.localPosition.dy).toInt());
-                x: (currentPosition.x + details.delta.dx).toInt(),
-                y: (currentPosition.y + details.delta.dy).toInt());
+            x:(currentPosition.x + details.delta.dx),
+            y:(currentPosition.y + details.delta.dy));
           });
           _notifyListeners();
         },
